backend/app/services/deploy_service.py
"""
Deployment service for Databricks jobs.

Placeholder implementation - returns mock job IDs. TODO: Integrate with Databricks SDK.
"""

from typing import Any

import uuid


class DeployService:
    """
    Service for deploying pipelines to Databricks as jobs.

    Placeholder implementation returns fake job IDs. Replace with actual
    Databricks SDK calls for production.
    """

    def deploy(
        self,
        pipeline_id: str,
        job_name: str,
        cluster_config: dict[str, Any],
        schedule: str | None,
        code: str,
    ) -> dict[str, str]:
        """
        Deploy a pipeline as a Databricks job.

        Args:
            pipeline_id: Pipeline identifier
            job_name: Name for the Databricks job
            cluster_config: Cluster configuration (instance type, workers, etc.)
            schedule: Optional cron schedule expression
            code: Generated notebook code to deploy

        Returns:
            Dict with job_id, job_url, and status

        TODO: Replace with actual Databricks SDK implementation:
            from databricks.sdk import WorkspaceClient
            w = WorkspaceClient()
            job = w.jobs.create(
                name=job_name,
                tasks=[{
                    "task_key": "main",
                    "notebook_task": {"notebook_path": "...", "source": "WORKSPACE"},
                    "job_cluster_key": "main_cluster",
                }],
                job_clusters=[{
                    "job_cluster_key": "main_cluster",
                    "new_cluster": cluster_config,
                }],
                schedule=({"quartz_cron_expression": schedule} if schedule else None),
            )
            return {"job_id": str(job.job_id), "job_url": job.url, "status": "created"}
        """
        # Mock implementation
        job_id = str(uuid.uuid4())
        job_url = f"https://databricks.example.com/#job/{job_id}"
        return {
            "job_id": job_id,
            "job_url": job_url,
            "status": "created",
        }

    def get_job_status(self, job_id: str) -> str:
        """
        Get the status of a deployed job.

        Args:
            job_id: Databricks job ID

        Returns:
            Job status string (e.g., "RUNNING", "TERMINATED", "PENDING")

        TODO: Replace with actual Databricks SDK implementation:
            from databricks.sdk import WorkspaceClient
            w = WorkspaceClient()
            run = w.jobs.get_run(job_id=job_id)
            return run.state.life_cycle_state
        """
        # Mock implementation
        return "PENDING"
