backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0
jinja2>=3.1
databricks-sdk>=0.38.0
python-dotenv
httpx
